Image Util PNG API
[Issue]    http://suprem.sec.samsung.net/jira/browse/TWF-1115
[Problem]  libpng was used, which is private now
[Solution] Replaced libpng with capi-media-image-util
[Verify]   Open few pages, check if thumbnails are saved in
           tab manager and in Quick Access

+++ core/Tools/Blob.cpp
@@ -41,7 +41,7 @@ Blob::Blob(Blob && other) : m_data(NULL), m_length(0)
     other.m_length = 0;
 }
 
-Blob::Blob(const void * ptr, int length) :
+Blob::Blob(const void* ptr, unsigned long long length) :
     m_length(length)
 {
     m_data = malloc(length);
@@ -85,7 +85,7 @@ const void * Blob::getData() const
     return this->m_data;
 }
 
-void Blob::setData(void * ptr, int length)
+void Blob::setData(void* ptr, unsigned long long length)
 {
     free();
